requirements.txt
amqp
asgiref
async-timeout==4.0.2
celery
celery-progress
certifi==2021.10.8
cffi
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django>=5.1
django-allauth==0.50.0
django-celery-results
django-debug-toolbar
gunicorn==20.1.0
idna==3.3
kombu
oauthlib==3.2.0
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
redis
requests==2.27.1
requests-oauthlib==1.3.1
sentry-sdk==1.5.12
six==1.16.0
spotipy>=2.24.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
vine
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
